requirements.txt
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0.1
pyarrow>=15.0.0
scipy>=1.12.0
numpy>=1.26.4
torch>=2.2.1
scikit-learn>=1.4.1.post1
statsmodels>=0.14.1
keras>=2.15.0
flask>=3.0.2